feat(runtime-c-api) Move the `wasi` module into its own file.
